feat: add codex cli provider prototype
