+++ charts/automate-e/templates/secret.yaml
@@ -9,5 +9,6 @@ type: Opaque
 stringData:
   discord-bot-token: {{ .Values.secrets.discordBotToken | quote }}
   anthropic-api-key: {{ .Values.secrets.anthropicApiKey | quote }}
+  openai-api-key: {{ .Values.secrets.openaiApiKey | quote }}
   database-url: {{ .Values.secrets.databaseUrl | quote }}
 {{- end }}
