+++ examples/example-e/character.json
@@ -54,8 +54,8 @@
     "format": "concise"
   },
   "llm": {
-    "provider": "anthropic",
-    "model": "claude-haiku-4-5-20251001",
+    "provider": "codex-cli",
+    "model": "gpt-5.4",
     "temperature": 0.7
   }
 }
