+++ examples/example-e/values.yaml
@@ -45,8 +45,8 @@ character:
     tone: "friendly and enthusiastic"
     format: concise
   llm:
-    provider: anthropic
-    model: claude-haiku-4-5-20251001
+    provider: codex-cli
+    model: gpt-5.4
     temperature: 0.7
 
 dashboard:
