+++ package.json
@@ -11,7 +11,7 @@
     "start:worker": "node src/worker.js",
     "dev": "node --watch src/index.js",
     "test": "node src/test.js",
-    "test:unit": "node --test src/conductor.test.js",
+    "test:unit": "node --test src/*.test.js",
     "run-once": "node src/run-once.js"
   },
   "dependencies": {
