+++ src/character.js
@@ -67,6 +67,7 @@ export function loadCharacter() {
   character.lore = character.lore || [];
   character.style = character.style || { language: 'English', tone: 'helpful', format: 'concise' };
   character.memory = character.memory || { conversationRetention: '30d' };
+  character.llm.provider = character.llm.provider || 'anthropic';
   character.llm.temperature = character.llm.temperature ?? 0.3;
   if (character.discord) {
     character.discord.allowBots = character.discord.allowBots || [];
